% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arial.R
\name{get_arial}
\alias{get_arial}
\title{Søg luftfotos}
\usage{
get_arial(param1, param2)
}
\arguments{
\item{param1}{Beskrivelse af første parameter.}

\item{param2}{Beskrivelse af anden parameter.}
}
\value{
Beskrivelse af returværdien.
}
\description{
En kort beskrivelse af hvad funktionen gør.
}
\examples{
# Eksempel på brug af funktionen
get_arial(param1 = 1, param2 = 2)
}
