% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_authors.R
\name{get_authors}
\alias{get_authors}
\title{Søg forfattere fra textcollections}
\usage{
get_authors(q)
}
\arguments{
\item{q}{søgestreng}
}
\value{
A tibble containing SVARET!
}
\description{
Hjælpe funktion til at returnere forfattere fra text collections
}
\examples{
# Eksempel på brug af funktionen
get_text(q = "helsinge")
}
