% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subcollection.R
\name{get_subcollections}
\alias{get_subcollections}
\title{Søg subcollections}
\usage{
get_subcollections()
}
\value{
A tibble containing SVARET!
}
\description{
returnerer subcollections fra text collections
}
\examples{
# Eksempel på brug af funktionen
get_text(q = "helsinge")
}
