% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_text}
\alias{get_text}
\title{Søg ADL}
\usage{
get_text(q)
}
\arguments{
\item{q}{søgestreng}
}
\value{
A tibble containing SVARET!
}
\description{
søger i ADL
}
\examples{
# Eksempel på brug af funktionen
get_text(q = "helsinge")
}
