% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KBapi-package.R
\docType{package}
\name{KBapi-package}
\alias{KBapi}
\alias{KBapi-package}
\title{KBapi: Wrapper for Royal Danish Library API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides wrappers for Royal Danish Library APIs for Aerial photographs, digitized images and litterature.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kubdatalab.github.io/KBAPI/}
}

}
\author{
\strong{Maintainer}: Christian Knudsen \email{cbk@kb.dk} (\href{https://orcid.org/0000-0002-9816-768X}{ORCID})

}
\keyword{internal}
