% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aerial.R
\name{get_aerial}
\alias{get_aerial}
\title{Søg luftfotos}
\usage{
get_aerial(lat = 56.0075146363177, lon = 12.2288401900055, format = "kml")
}
\arguments{
\item{lat}{Latitude}

\item{lon}{Longitude}

\item{format}{Outputformat. Either kml, rss, atom or mods}
}
\value{
Beskrivelse af returværdien.
}
\description{
En kort beskrivelse af hvad funktionen gør.
}
\examples{
# Eksempel på brug af funktionen
get_aerial(lat = 1, lon = 2)
}
