% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categories.R
\name{get_categories}
\alias{get_categories}
\title{Categories in text collections}
\usage{
get_categories()
}
\value{
A character vector containing current categories
}
\description{
returnerer kategorier fra text collections
}
\examples{
# Eksempel på brug af funktionen
get_categories()
}
