% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subcollection.R
\name{get_subcollections}
\alias{get_subcollections}
\title{Get sub collections}
\usage{
get_subcollections()
}
\value{
A character vector containing current list of sub collections
}
\description{
returns sub collections from text collections
}
\examples{
# Eksempel på brug af funktionen
get_subcollections()
}
